+++ pkg/controller/storagecluster/external_resources_test.go
@@ -1,0 +1,175 @@
+package storagecluster
+
+import (
+	"encoding/json"
+	"fmt"
+	nbv1 "github.com/noobaa/noobaa-operator/v2/pkg/apis/noobaa/v1alpha1"
+	api "github.com/openshift/ocs-operator/pkg/apis/ocs/v1"
+	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
+	storagev1 "k8s.io/api/storage/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	"testing"
+)
+
+var ExternalResources = []ExternalResource{
+	ExternalResource{
+		Kind: "ConfigMap",
+		Data: map[string]string{
+			"maxMonId": "0",
+			"data":     "a=10.20.30.40:1234",
+			"mapping":  "{}",
+		},
+		Name: "rook-ceph-mon-endpoints",
+	},
+	ExternalResource{
+		Kind: "Secret",
+		Data: map[string]string{
+			"userKey": "someUserKeyRBD==",
+			"userID":  "csi-rbd-node",
+		},
+		Name: "rook-csi-rbd-node",
+	},
+	ExternalResource{
+		Kind: "StorageClass",
+		Data: map[string]string{
+			"pool": "device_health_metrics",
+		},
+		Name: "ceph-rbd",
+	},
+	ExternalResource{
+		Kind: "StorageClass",
+		Data: map[string]string{
+			"fsName": "myfs",
+			"pool":   "myfs-data0",
+		},
+		Name: "cephfs",
+	},
+	ExternalResource{
+		Kind: "StorageClass",
+		Data: map[string]string{
+			"endpoint": "10.20.30.40:50",
+		},
+		Name: "ceph-rgw",
+	},
+}
+
+func TestEnsureExternalStorageClusterResources(t *testing.T) {
+	request := reconcile.Request{
+		NamespacedName: types.NamespacedName{
+			Name:      "ocsinit",
+			Namespace: "",
+		},
+	}
+	reconciler := createExternalClusterReconciler(t)
+	result, err := reconciler.Reconcile(request)
+	assert.NoError(t, err)
+	assert.Equal(t, reconcile.Result{}, result)
+	assertExpectedExternalResources(t, reconciler)
+}
+
+func createExternalCephClusterSecret() (*corev1.Secret, error) {
+	jsonBlob, err := json.Marshal(ExternalResources)
+	if err != nil {
+		return nil, err
+	}
+	externalSecret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: externalClusterDetailsSecret,
+		},
+		Data: map[string][]byte{
+			externalClusterDetailsKey: jsonBlob,
+		},
+	}
+	return externalSecret, err
+}
+
+func createExternalClusterReconciler(t *testing.T) ReconcileStorageCluster {
+	cr := &api.StorageCluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "ocsinit",
+		},
+		Spec: api.StorageClusterSpec{
+			ExternalStorage: api.ExternalStorageClusterSpec{
+				Enable: true,
+			},
+		},
+	}
+	externalSecret, err := createExternalCephClusterSecret()
+	if err != nil {
+		t.Fatalf("failed to create external secret: %v", err)
+	}
+	reconciler := createFakeInitializationStorageClusterReconciler(t, &nbv1.NooBaa{})
+	runtimeObjs := []runtime.Object{cr, externalSecret}
+	for _, obj := range runtimeObjs {
+		if err = reconciler.client.Create(nil, obj); err != nil {
+			t.Fatalf("failed to create a needed runtime object: %v", err)
+		}
+	}
+	return reconciler
+}
+
+func assertExpectedExternalResources(t assert.TestingT, reconciler ReconcileStorageCluster) {
+	request := reconcile.Request{
+		NamespacedName: types.NamespacedName{
+			Name:      "ocsinit",
+			Namespace: "",
+		},
+	}
+
+	sc := &api.StorageCluster{}
+	err := reconciler.client.Get(nil, request.NamespacedName, sc)
+	assert.NoError(t, err)
+
+	externalSecret := &corev1.Secret{}
+	request.Name = externalClusterDetailsSecret
+	err = reconciler.client.Get(nil, request.NamespacedName, externalSecret)
+	assert.NoError(t, err)
+
+	var data []ExternalResource
+	err = json.Unmarshal(externalSecret.Data[externalClusterDetailsKey], &data)
+	if err != nil {
+		t.Errorf("fatal err %+v", err)
+	}
+
+	for _, expected := range data {
+		request.Name = expected.Name
+		switch expected.Kind {
+		case "ConfigMap":
+			actual := &corev1.ConfigMap{}
+			err := reconciler.client.Get(nil, request.NamespacedName, actual)
+			assert.NoError(t, err)
+			for er := range expected.Data {
+				assert.Equal(t, expected.Data[er], actual.Data[er])
+			}
+		case "Secret":
+			actual := &corev1.Secret{}
+			err := reconciler.client.Get(nil, request.NamespacedName, actual)
+			assert.NoError(t, err)
+			for er := range expected.Data {
+				assert.Equal(t, []byte(expected.Data[er]), actual.Data[er])
+			}
+		case "StorageClass":
+			actual := &storagev1.StorageClass{}
+			request.Name = fmt.Sprintf("%s-%s", sc.Name, expected.Name)
+			err := reconciler.client.Get(nil, request.NamespacedName, actual)
+			assert.NoError(t, err)
+			for param, value := range expected.Data {
+				assert.Equal(t, value, actual.Parameters[param])
+			}
+			// Verify the RGW SC parameters in external mode are correct
+			// The main difference between external and converged is the presence of an endpoint
+			// and the absence of the "objectStoreName" parameter
+			if actual.Name == "ocsinit-ceph-rgw" {
+				assert.NotEmpty(t, actual.Parameters["endpoint"], actual.Parameters["endpoint"])
+				assert.NotEmpty(t, actual.Parameters["region"], actual.Parameters)
+				assert.NotContains(t, actual.Parameters["objectStoreName"], actual.Parameters)
+				assert.Equal(t, actual.Parameters["region"], "us-east-1")
+				assert.Equal(t, 3, len(actual.Parameters), actual.Parameters)
+			}
+		}
+	}
+}
